Delay update 1. Fixed delay on reply message 2. Update /help command's text 3. Added message delay in group for /start, /help 4. Function answer and reply in funcs.py
