+++ config/languages-json/UKR.json
@@ -6,7 +6,7 @@
 
     "start": " 👋 Вітаю в нашій когорті, хлопче/дівчино!\nЦей ботик допоможе тобі файно провести вільний час, граючись із розмірами твоїх ягідок. Скоріше приймай участь, buddy, та не зволікай, пиши /ass",
 
-    "help": "Не бійсь, ось усі команди:\n----------------------\n🍑  /ass — почати гру\n🎰 /luck — казино\n❔  /help — ця довідка\n🚶‍♂️  /leave — покинути гру\n🙋‍♂️ /r повністю видали російську мову — повідомити про недолік або ідею\n📊 /statistic — рейтинг гравців\n\n 👨‍💻 /about — про розробника\n ----------------------",
+    "help": "Не бійсь, ось усі команди:\n----------------------\n🍑  /ass — почати гру\n🎰 /luck — казино\n❔  /help — ця довідка\n🚶‍♂️  /leave — покинути гру\n🙋‍♂️ /r повністю видали російську мову — повідомити про недолік або ідею\n📊 /statistic — рейтинг гравців\n\n 👨‍💻 /about — про розробника\n ----------------------\nЯкщо ви бажаєте, щоб ботик видаляв команди користувачів після обробки - зробіть ботика адміном ♥",
 
 	"admin": {
 		"help": "Режим баті:\n----------------------\n Замість аргументів можна використовувати параметр 'self'\n /bl group_id — показати заблокованих гравців\n /ub group_id user_id — розбанити гравця\n /ban group_id user_id — забанити покидька\n /show group_id — список гравція за групою\n /reports — показати звіти\n /dreports — показати детальні звіти\n /groups — показати усі групи\n /clear — видалити усі звіти\n /notify — віправити повідомлення группам\n ----------------------",
